CartCove/requirements.txt
asgiref==3.7.2
black==23.10.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==5.0.1
Django==4.2
django-allauth==0.58.0
django-cors-headers==4.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==20.1.0
filelock==3.12.4
identify==2.5.31
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
pipenv==2023.10.3
platformdirs==3.11.0
pre-commit==3.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
spark-parser==1.8.9
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uncompyle6==3.9.0
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.5
xdis==6.0.5
